Large update, config is now one yaml file which is much easier to read and allows more complexity
There is much more fine-tuning possible when it comes to similarity metrics, a different metric can be chosen and tuned per predicate. Possible metrics at this point are jarowinkler, ngram and token.

+++ src/main/java/org/uu/nl/embedding/glove/opt/GloveJob.java
@@ -5,4 +5,4 @@ import java.util.concurrent.Callable;
 /**
  * @author Jurian Baas
  */
-public interface GloveJob extends Callable<Double> { }
+public interface GloveJob extends Callable<Float> { }
